requirements.txt
requests==2.26.0
pyyaml==6.0
pytest==2.7.2
isort==5.9.3
flake8==4.0.1
virtualenv==20.8.1
urllib3==1.26.7
responses==0.14.0
tenacity==8.0.1